feat(bridge): Introduce a settings sub-navigation (#4501)
