+++ bridge/client/app/_components/ktb-project-settings/ktb-project-settings.component.scss
