+++ bridge/client/app/_views/ktb-uniform-view/ktb-uniform-view.component.html
@@ -1,11 +1,11 @@
 <div fxFlexFill>
   <div fxFlex fxLayout="row">
-    <dt-menu class="uniform-submenu" aria-label="Keptn Uniform Submenu" fxFlex="0 0 140px">
+    <dt-menu class="submenu" aria-label="Keptn Uniform Submenu" fxFlex="0 0 200px">
       <dt-menu-group>
-        <button dt-icon-button dtMenuItem variant="nested" aria-label="Open environment view" routerLinkActive="active" routerLink="services">
+        <button dtMenuItem variant="nested" aria-label="Open uniform services" routerLinkActive="active" routerLink="services">
           Uniform
         </button>
-        <button dt-icon-button dtMenuItem variant="nested" aria-label="Open services view" routerLinkActive="active" routerLink="secrets">
+        <button dtMenuItem variant="nested" aria-label="Open uniform secrets" routerLinkActive="active" routerLink="secrets">
           Secrets
         </button>
       </dt-menu-group>
