+++ bridge/client/app/project-board/project-board.component.html
@@ -2,23 +2,35 @@
   <dt-empty-state>
     <dt-empty-state-item>
       <dt-empty-state-item-img>
-        <img alt="Keptn logo" class="mt-2" [src]="logoInvertedUrl" />
+        <img alt="Keptn logo" class="mt-2" [src]="logoInvertedUrl"/>
       </dt-empty-state-item-img>
       <ng-container [ngSwitch]="error">
         <ng-container *ngSwitchCase="'project'">
           <dt-empty-state-item-title class="mt-2" aria-level="2">Project doesn't exist</dt-empty-state-item-title>
-          <p>Sorry, the project could not be loaded. Check out the <a href="https://keptn.sh/docs/quickstart/" target="_blank" rel="noopener noreferrer">Quick Start</a> instructions on how to <a [href]="'/manage/project/' | keptnUrl" target="_blank">create a project</a> or go back to <a [routerLink]="['/dashboard']">your dashboard</a>.</p>
-          <button dt-button (click)="loadProjects()"><dt-icon name="refresh"></dt-icon> Retry</button>
+          <p>Sorry, the project could not be loaded. Check out the <a href="https://keptn.sh/docs/quickstart/" target="_blank" rel="noopener noreferrer">Quick Start</a> instructions on how to <a [href]="'/manage/project/' | keptnUrl" target="_blank">create a project</a> or go back to <a
+            [routerLink]="['/dashboard']">your dashboard</a>.</p>
+          <button dt-button (click)="loadProjects()">
+            <dt-icon name="refresh"></dt-icon>
+            Retry
+          </button>
         </ng-container>
         <ng-container *ngSwitchCase="'trace'">
           <dt-empty-state-item-title class="mt-2" aria-level="2">Traces for <span class="italic" [textContent]="contextId"></span> not found</dt-empty-state-item-title>
-          <p>Sorry, traces with this shkeptncontext could not be loaded. Check out the <a href="https://keptn.sh/docs/quickstart/" target="_blank" rel="noopener noreferrer">Quick Start</a> instructions on how to <a [href]="'/manage/project/' | keptnUrl" target="_blank">create a project</a> or go back to <a [routerLink]="['/dashboard']">your dashboard</a>.</p>
-          <button dt-button (click)="loadProjects()"><dt-icon name="refresh"></dt-icon> Retry</button>
+          <p>Sorry, traces with this shkeptncontext could not be loaded. Check out the <a href="https://keptn.sh/docs/quickstart/" target="_blank" rel="noopener noreferrer">Quick Start</a> instructions on how to <a [href]="'/manage/project/' | keptnUrl" target="_blank">create a project</a> or go
+            back to <a [routerLink]="['/dashboard']">your dashboard</a>.</p>
+          <button dt-button (click)="loadProjects()">
+            <dt-icon name="refresh"></dt-icon>
+            Retry
+          </button>
         </ng-container>
         <ng-container *ngSwitchDefault>
           <dt-empty-state-item-title class="mt-2" aria-level="2">Some error occurred</dt-empty-state-item-title>
-          <p>Sorry, some error occurred. Check out the <a href="https://keptn.sh/docs/quickstart/" target="_blank" rel="noopener noreferrer">Quick Start</a> instructions on how to <a [href]="'/manage/project/' | keptnUrl" target="_blank">create a project</a> or go back to <a [routerLink]="['/dashboard']">your dashboard</a>.</p>
-          <button dt-button (click)="loadProjects()"><dt-icon name="refresh"></dt-icon> Retry</button>
+          <p>Sorry, some error occurred. Check out the <a href="https://keptn.sh/docs/quickstart/" target="_blank" rel="noopener noreferrer">Quick Start</a> instructions on how to <a [href]="'/manage/project/' | keptnUrl" target="_blank">create a project</a> or go back to <a
+            [routerLink]="['/dashboard']">your dashboard</a>.</p>
+          <button dt-button (click)="loadProjects()">
+            <dt-icon name="refresh"></dt-icon>
+            Retry
+          </button>
         </ng-container>
       </ng-container>
     </dt-empty-state-item>
@@ -41,11 +53,11 @@
 
 
 <ng-template #keptnMenu let-isCreateMode="isCreateMode">
-  <dt-menu aria-label="Keptn Menu" class="dt-default-menu-example" fxFlex="0 0 50px">
+  <dt-menu aria-label="Keptn Menu" class="main-nav" fxFlex="0 0 50px">
     <dt-menu-group>
       <ng-template #overlay let-data><span>{{data.text}}</span></ng-template>
 
-      <button dt-icon-button dtMenuItem variant="nested" aria-label="Open environment view" [disabled]="isCreateMode" [dtOverlay]="overlay" [dtOverlayConfig]="{data: {text: 'Environment'}}" routerLinkActive="active" routerLink="./" [routerLinkActiveOptions]= "{exact: true}">
+      <button dt-icon-button dtMenuItem variant="nested" aria-label="Open environment view" [disabled]="isCreateMode" [dtOverlay]="overlay" [dtOverlayConfig]="{data: {text: 'Environment'}}" routerLinkActive="active" routerLink="./" [routerLinkActiveOptions]="{exact: true}">
         <dt-icon name="environment"></dt-icon>
       </button>
