+++ bridge/client/styles.scss
@@ -137,6 +137,22 @@ button.dt-button-secondary.error-button:not([disabled]) {
   }
 }
 
+.dt-menu.submenu .dt-menu-group:first-of-type {
+  margin-top: 15px;
+}
+
+.dt-menu.submenu .dt-menu-item.active,
+.dt-menu.submenu .dt-menu-item:hover,
+.dt-menu.submenu .dt-menu-item:active {
+  background-color: $gray-200;
+}
+
+.dt-menu.submenu,
+.dt-menu.submenu .dt-menu-group-header,
+.dt-menu.submenu .dt-menu-item {
+  color: $default-font-color;
+  background-color: $gray-100;
+}
 
 .m-0 {
   margin: 0 !important;
